sync on 2026/06/13, rev 12ee3f21d4555dd6a142986ef3d38645f1c0a33e

+++ prog/daNetGame/net/locSnapshots.h
@@ -11,7 +11,7 @@ struct LocSnapshot
 {
   Quat quat;
   Point3 pos;
-  float atTime;
+  double atTime;
   float interval;
   bool blink;
